CI: Workflow improvement

+++ pytoda/datasets/tests/test_set_matching_dataset.py
@@ -87,8 +87,8 @@ class TestSetMatchingDataset(unittest.TestCase):
 
                                 self.assertTrue(
                                     torch.equal(
-                                        sample1[1][sample1[2], :],
-                                        sample2[1][sample2[2], :],
+                                        sample1[1][sample1[2].long(), :],
+                                        sample2[1][sample2[2].long(), :],
                                     )
                                 )
 
@@ -98,13 +98,13 @@ class TestSetMatchingDataset(unittest.TestCase):
 
                                     self.assertTrue(
                                         torch.equal(
-                                            sample1[1][sample1[2], :],
+                                            sample1[1][sample1[2].long(), :],
                                             sample1[0],
                                         )
                                     )
                                     self.assertTrue(
                                         torch.equal(
-                                            sample1[0][sample1[3], :],
+                                            sample1[0][sample1[3].long(), :],
                                             sample1[1],
                                         )
                                     )
@@ -113,13 +113,13 @@ class TestSetMatchingDataset(unittest.TestCase):
 
                                     self.assertFalse(
                                         torch.equal(
-                                            sample1[1][sample1[2], :],
+                                            sample1[1][sample1[2].long(), :],
                                             sample1[0],
                                         )
                                     )
                                     self.assertFalse(
                                         torch.equal(
-                                            sample1[0][sample1[3], :],
+                                            sample1[0][sample1[3].long(), :],
                                             sample1[1],
                                         )
                                     )
@@ -142,8 +142,8 @@ class TestSetMatchingDataset(unittest.TestCase):
                                     if noise == 0.0:
                                         self.assertTrue(
                                             torch.equal(
-                                                sample1[1][sample1[2], :],
-                                                sample2[1][sample2[2], :],
+                                                sample1[1][sample1[2].long(), :],
+                                                sample2[1][sample2[2].long(), :],
                                             )
                                         )
                                 elif sample1[-1] != sample2[-1]:
